Render Meshlets (#29)
* feat: Uniform testing

* feat: Draw Meshlets

* chore: Added stricter compiler flags and fixed all warnings

* feat: Draw Cube

* feat: constant buffer instead of root constants

* feat: Packed Buffer

* feat: Testing Framework

* chore: Add noexcept

* feat: Meshlet deletion

+++ bgl/src/d3d12/ErrorChecker.cpp
@@ -7,14 +7,14 @@ namespace bgl
 	{
 		wchar_t*   descriptionWinalloc = nullptr;
 		const auto result              = FormatMessageW(
-            FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM |
-                FORMAT_MESSAGE_IGNORE_INSERTS,
-            nullptr,
-            hr,
-            MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
-            reinterpret_cast<LPWSTR>(&descriptionWinalloc),
-            0,
-            nullptr);
+			FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM |
+				FORMAT_MESSAGE_IGNORE_INSERTS,
+			nullptr,
+			static_cast<DWORD>(hr),
+			MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
+			reinterpret_cast<LPWSTR>(&descriptionWinalloc),
+			0,
+			nullptr);
 
 		std::wstring description;
 		if (result)
